bridge/game 改写
增加util/UUID模块
补全了一些命名空间

+++ minecraft/network/chat/TextColor.h
@@ -40,7 +40,7 @@ namespace cerver {
 
                     std::string formatValue() const;
 
-                    int hashCode();
+                    size_t hashCode();
 
                     std::string toString();
